Move reclaimer guard to velox/common
